fix is_allowed_named_entity type

+++ src/html_entities.h
@@ -10,7 +10,7 @@ const size_t MAX_NUM_ENTITY_LEN;
 
 int is_valid_numeric_entity(uint32_t entity_val);
 
-const char* is_allowed_named_entity (register const char *str, register unsigned int len);
+const char* is_allowed_named_entity (register const char *str, register size_t len);
 
 #endif
